move general-ish kubeconfig package internal to cluster kubeconfig management
