+++ pkg/internal/cluster/kubeconfig/internal/kubeconfig/paths.go
