+++ pkg/internal/cluster/kubeconfig/internal/kubeconfig/paths_test.go
