+++ pkg/internal/cluster/kubeconfig/internal/kubeconfig/remove.go
