+++ pkg/internal/cluster/kubeconfig/internal/kubeconfig/remove_test.go
